refactor: :recycle: выделить фазы project validation
